Fixed mess after merge
Compiler was confusing `App.Models.Task` with `System.Threading.Task`. I tried to find a global using exporting `Threading`, but found none; Used a fully qualified alias `using Task = App.Models.Task` to resolve the issue.

Also updated `TaskController` to use `long` for `task_id` instead of `int`.

+++ App/Controllers/TaskController.cs
@@ -41,7 +41,7 @@ public class TaskController
         return new Task(id, title, description, created, due, false);
     }
 
-    public async System.Threading.Tasks.Task CompleteTask(int taskId)
+    public async System.Threading.Tasks.Task CompleteTask(long taskId)
     {
         await System.Threading.Tasks.Task.Delay(10);
     }
